+++ app/controllers/konacha/specs_controller.rb
@@ -4,8 +4,13 @@ module Konacha
       render :text => "Not found", :status => 404
     end
 
+    def reporter
+      @iframe_path = runner_path + request.fullpath
+    end
+
     def specs
       @specs = Konacha::Spec.find(params[:path] || "")
+      @stylesheets = Konacha::Engine.config.konacha.stylesheets
     end
   end
 end